+++ Firmware/temperature.cpp
@@ -38,7 +38,7 @@
 #include "temperature.h"
 #include "cardreader.h"
 
-#include "Sd2PinMap.h"
+#include "SdFatUtil.h"
 
 #include <avr/wdt.h>
 #include "adc.h"
@@ -2064,6 +2064,9 @@ FORCE_INLINE static void temperature_isr()
     }
   }
 #endif //BABYSTEPPING
+
+  // Check if a stack overflow happened
+  if (!SdFatUtil::test_stack_integrity()) stack_error();
 
 #if (defined(FANCHECK) && defined(TACH_0) && (TACH_0 > -1))
   check_fans();
